Add realtime message appending
