+++ app/modules/app.js
@@ -1,9 +1,8 @@
 import {getItem} from '../utils/storage'
 import {getCurrentUser} from './viewer'
-import {getRooms, getSuggestedRooms, subscribeToRooms, updateRoomState} from './rooms'
-import {appendMessages} from './messages'
-import FayeGitter from '../../libs/react-native-gitter-faye'
-import {DeviceEventEmitter, NetInfo, AppState} from 'react-native'
+import {getRooms, getSuggestedRooms} from './rooms'
+import {NetInfo, AppState} from 'react-native'
+import {setupFayeEvents, setupFaye, onNetStatusChangeFaye} from './realtime'
 
 /**
  * Constants
@@ -12,7 +11,6 @@ import {DeviceEventEmitter, NetInfo, AppState} from 'react-native'
 export const INITIALIZED = 'app/INITIALIZED'
 export const CHANGE_NET_STATUS = 'app/CHANGE_NET_STATUS'
 export const CHANGE_APP_STATE = 'app/CHANGE_APP_STATE'
-export const FAYE_CONNECT = 'app/FAYE_CONNECT'
 
 /**
  * Action Creators
@@ -24,8 +22,8 @@ export function init() {
     dispatch(setupFayeEvents())
     try {
       const token = await getItem('token')
-      const netStatus = await NetInfo.fetch()
-      dispatch({ type: INITIALIZED, token, netStatus })
+      // const netStatus = await NetInfo.fetch()
+      dispatch({ type: INITIALIZED, token })
 
       // TODO: do things belowe only if the internet is awailible (netStatus)
 
@@ -45,33 +43,6 @@ export function init() {
   }
 }
 
-function setupFaye() {
-  return async (dispatch, getState) => {
-    FayeGitter.setAccessToken(getState().auth.token)
-    FayeGitter.create()
-    FayeGitter.logger()
-    try {
-      const result = await FayeGitter.connect()
-      dispatch({type: FAYE_CONNECT, payload: result})
-      dispatch(subscribeToRooms())
-    } catch (err) {
-      console.log(err) // eslint-disable-line no-console
-    }
-  }
-}
-
-function onNetStatusChangeFaye(status) {
-  return async (dispatch, getState) => {
-    const {fayeConnected} = getState().app
-    if (!status && fayeConnected) {
-      dispatch({type: FAYE_CONNECT, payload: status})
-    }
-    if (status && !fayeConnected) {
-      dispatch(setupFaye())
-    }
-  }
-}
-
 function setupNetStatusListener() {
   return dispatch => {
     NetInfo.isConnected.addEventListener('change',
@@ -84,59 +55,12 @@ function setupNetStatusListener() {
 }
 
 function setupAppStatusListener() {
-  return dispatch => {
-    AppState.addEventListener('change',
-      status => dispatch({type: CHANGE_APP_STATE, payload: status})
-    );
-  }
-}
-
-
-function setupFayeEvents() {
   return (dispatch, getState) => {
-    DeviceEventEmitter
-      .addListener('FayeGitter:onDisconnected', log => {
-        console.warn(log) // eslint-disable-line no-console
-        dispatch(setupFaye())
-      })
-
-    DeviceEventEmitter
-      .addListener('FayeGitter:onFailedToCreate', log => {
-        console.warn(log) // eslint-disable-line no-console
-        if (getState().app.netStatus) {
-          dispatch(setupFaye())
-        }
-      })
-    DeviceEventEmitter
-      .addListener('FayeGitter:Message', event => {
-        dispatch(parseEvent(event))
-      })
-    DeviceEventEmitter
-      .addListener('FayeGitter:SubscribtionFailed', log => console.warn(log)) // eslint-disable-line no-console
-    DeviceEventEmitter
-      .addListener('FayeGitter:Subscribed', log => console.log(log)) // eslint-disable-line no-console
-    DeviceEventEmitter
-      .addListener('FayeGitter:Unsubscribed', log => console.log(log)) // eslint-disable-line no-console
-    DeviceEventEmitter
-      .addListener('FayeGitter:log', log => console.log(log)) // eslint-disable-line no-console
-  }
-}
-
-function parseEvent(event) {
-  return (dispatch, getState) => {
-    const message = JSON.parse(event.json)
-    const {id} = getState().viewer.user
-    const {activeRoom} = getState().rooms
-    const roomsChannel = `/api/v1/user/${id}/rooms`
-    const chatMessages = `/api/v1/rooms/${activeRoom}/chatMessages`
-
-    if (event.channel.match(roomsChannel)) {
-      dispatch(updateRoomState(message))
-    }
-
-    if (event.channel.match(chatMessages)) {
-      dispatch(appendMessages(activeRoom, [message.model]))
-    }
+    AppState.addEventListener('change', status => {
+      // TODO: Update drawer rooms state and messages in current room
+      // if app status changes from backgrount to active
+      dispatch({type: CHANGE_APP_STATE, payload: status})
+    })
   }
 }
 
@@ -146,7 +70,7 @@ function parseEvent(event) {
  */
 
 const initialState = {
-  online: null,
+  online: false,
   appState: null,
   fayeConnected: false
 }
@@ -154,9 +78,10 @@ const initialState = {
 export default function app(state = initialState, action) {
   switch (action.type) {
   case INITIALIZED:
-    return {...state,
-      online: action.netStatus
-    }
+    // return {...state,
+    //   online: action.netStatus
+    // }
+    return state
 
   case CHANGE_NET_STATUS:
     return {...state,
@@ -166,11 +91,6 @@ export default function app(state = initialState, action) {
   case CHANGE_APP_STATE:
     return {...state,
       appState: action.payload
-    }
-
-  case FAYE_CONNECT:
-    return {...state,
-      fayeConnected: action.payload
     }
 
   default:
