+++ app/screens/RoomScreen.js
@@ -21,7 +21,6 @@ import {
   prepareListView,
   getRoomMessagesBefore,
   getRoomMessagesIfNeeded,
-  subscribeToChatMessages,
   sendMessage,
   resendMessage,
   updateMessage
